Module_12/book_service/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
confluent-kafka
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-sqlalchemy
opentelemetry-exporter-zipkin
opentelemetry-exporter-prometheus
prometheus-client

structlog
opentelemetry-instrumentation-logging
python-json-logger